# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Граница API",
        "version": "1.0.0"
    },
    "paths": {
        "/api/directions": {
            "get": {
                "tags": [
                    "Direction"
                ],
                "summary": "Получение всех направлений",
                "operationId": "a6ce81d165ecb953dda16b03c6722076",
                "responses": {
                    "200": {
                        "description": "Request Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Direction"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/directions/borderCrossing": {
            "get": {
                "tags": [
                    "BorderCrossing"
                ],
                "summary": "Получение всех погран-переходов по направлению",
                "operationId": "a866e3b9e43b2e8ee5ae78242f05a9e1",
                "responses": {
                    "200": {
                        "description": "Request Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/BorderCrossing"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка при не переданном GET парметре",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Не передан directionId"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/directions/borderCrossing/cameras": {
            "get": {
                "tags": [
                    "Cameras"
                ],
                "summary": "Получение всех камер по погран-переходу",
                "operationId": "43cae11791cb64c7042abed8a634bba3",
                "responses": {
                    "200": {
                        "description": "Request Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Camera"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка при не переданном GET парметре",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Не передан borderCrossingId"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/directions/borderCrossing/reports/last": {
            "get": {
                "tags": [
                    "Report"
                ],
                "summary": "Получение последних 6-и отчетов погран-перехода",
                "operationId": "bd5975c8693ec3b89b12a950f94c2474",
                "responses": {
                    "200": {
                        "description": "Request Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/LastReportDTO"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка при не переданном GET парметре",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Не передан borderCrossingId"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/directions/borderCrossing/reports": {
            "get": {
                "tags": [
                    "Report"
                ],
                "summary": "Получение всех отчетов погран-перехода",
                "operationId": "650562ca1644f2c5585234f6f270dcc6",
                "responses": {
                    "200": {
                        "description": "Request Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/AllReportDTO"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Ошибка при не переданном GET парметре",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Не передан borderCrossingId"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Report"
                ],
                "summary": "Создать новый отчет",
                "operationId": "2fa6b45bbd6bc7430bf030cae4c41696",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "border_crossing_id",
                                    "transport_id",
                                    "user_id",
                                    "checkpoint_entry",
                                    "checkpoint_exit"
                                ],
                                "properties": {
                                    "border_crossing_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "transport_id": {
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "user_id": {
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "checkpoint_queue": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-07-01T08:00:00Z"
                                    },
                                    "checkpoint_entry": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-07-01T09:00:00Z"
                                    },
                                    "checkpoint_exit": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-07-01T10:00:00Z"
                                    },
                                    "comment": {
                                        "type": "string",
                                        "example": "Sample comment"
                                    },
                                    "is_flipped_direction": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "time_enter_waiting_area": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-07-01T10:00:00Z"
                                    },
                                    "time_leave_waiting_area": {
                                        "type": "string",
                                        "format": "date-time",
                                        "example": "2024-07-01T10:00:00Z"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Report Created Successfully"
                    },
                    "422": {
                        "description": "Ошибка если Entity не прошел валидацию",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The transport id field is required."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/directions/borderCrossing/reports/transports": {
            "get": {
                "tags": [
                    "Transport"
                ],
                "summary": "Получение всех видов ТС",
                "operationId": "ee8bff5a672458a4ed2482a8bcd0e905",
                "responses": {
                    "200": {
                        "description": "Request Successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Transport"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Direction": {
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "North"
                    },
                    "logo": {
                        "type": "string",
                        "example": "path to logo"
                    },
                    "image": {
                        "type": "string",
                        "example": "path to image"
                    },
                    "info": {
                        "type": "string",
                        "example": "info"
                    }
                },
                "type": "object"
            },
            "BorderCrossing": {
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "is_queue": {
                        "type": "boolean",
                        "example": true
                    },
                    "header_image": {
                        "type": "string",
                        "example": "test.png"
                    },
                    "from_city": {
                        "properties": {
                            "name": {
                                "type": "string",
                                "example": "City1"
                            },
                            "country": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Country1"
                                    },
                                    "logo": {
                                        "type": "string",
                                        "example": "img.png"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "to_city": {
                        "properties": {
                            "name": {
                                "type": "string",
                                "example": "City2"
                            },
                            "country": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Country2"
                                    },
                                    "logo": {
                                        "type": "string",
                                        "example": "img.png"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "type": "object"
                    },
                    "url_arcticle": {
                        "type": "string",
                        "example": "www.google.com"
                    },
                    "cache": {
                        "properties": {
                            "time": {
                                "type": "string",
                                "example": "2024-01-01 14:00:00"
                            },
                            "countCar": {
                                "type": "integer",
                                "example": "6"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "Camera": {
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "url": {
                        "type": "string",
                        "example": "https://www.camera.ru"
                    },
                    "description": {
                        "type": "string",
                        "example": "Desc"
                    },
                    "photo": {
                        "type": "string",
                        "example": "photo.png"
                    }
                },
                "type": "object"
            },
            "LastReportDTO": {
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "checkpoint_queue": {
                        "type": "string",
                        "example": "2024-08-14 13:04:00"
                    },
                    "checkpoint_entry": {
                        "type": "string",
                        "example": "2024-08-14 13:04:00"
                    },
                    "checkpoint_exit": {
                        "type": "string",
                        "example": "2024-08-14 13:04:00"
                    },
                    "comment": {
                        "type": "string",
                        "example": "test"
                    },
                    "is_flipped_direction": {
                        "type": "boolean",
                        "example": true
                    },
                    "is_show_button": {
                        "type": "boolean",
                        "example": "true"
                    }
                },
                "type": "object"
            },
            "AllReportDTO": {
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "checkpoint_queue": {
                        "type": "string",
                        "example": "2024-08-14 13:04:00"
                    },
                    "checkpoint_entry": {
                        "type": "string",
                        "example": "2024-08-14 13:04:00"
                    },
                    "checkpoint_exit": {
                        "type": "string",
                        "example": "2024-08-14 13:04:00"
                    },
                    "comment": {
                        "type": "string",
                        "example": "test"
                    },
                    "is_flipped_direction": {
                        "type": "boolean",
                        "example": true
                    },
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "transport": {
                        "properties": {
                            "icon": {
                                "type": "string",
                                "example": "img.png"
                            }
                        },
                        "type": "object"
                    },
                    "time_enter_waiting_area": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-01T10:00:00Z"
                    },
                    "time_leave_waiting_area": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-07-01T10:00:00Z"
                    },
                    "time_difference_text": {
                        "type": "string",
                        "example": "11 часов 6 минут"
                    }
                },
                "type": "object"
            },
            "Transport": {
                "required": [
                    "id",
                    "name"
                ],
                "properties": {
                    "icon": {
                        "type": "string",
                        "example": "img.png"
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Direction",
            "description": "Direction"
        },
        {
            "name": "BorderCrossing",
            "description": "BorderCrossing"
        },
        {
            "name": "Cameras",
            "description": "Cameras"
        },
        {
            "name": "Report",
            "description": "Report"
        },
        {
            "name": "Transport",
            "description": "Transport"
        }
    ]
}